+++ SystemControl/nid_resolver.c
@@ -98,7 +98,7 @@ u32 resolve_nid(resolver_config *resolver, u32 nid)
 		new = resolver->nidtable[i].new;
 
 		if(new != UNKNOWNNID) {
-			printk("%s: %s_%08X->%s_%08X\n", __func__, resolver->name, nid, resolver->name, new);
+			printk("%s: %s_%08X->%s_%08X\n", __func__, resolver->name, (uint)nid, resolver->name, (uint)new);
 
 			return new;
 		}
@@ -130,7 +130,7 @@ static int resolve_missing_nid(SceLibraryStubTable *stub, MissingNIDResolver *re
 
 				stub_addr = stub->stubtable + (i << 3);
 				fp = resolver->entry[j].fp;
-				printk("%s: %s_%08X resolved(fp: 0x%08X)\n", __func__, libname, nid, fp);
+				printk("%s: %s_%08X resolved(fp: 0x%08X)\n", __func__, libname, (uint)nid, (uint)fp);
 				REDIRECT_FUNCTION(fp, (u32)stub_addr);
 				sceKernelDcacheWritebackInvalidateRange(stub_addr, 8);
 				sceKernelIcacheInvalidateRange(stub_addr, 8);
